# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bio?: string | null;
  /**
     * JSON-encoded personality traits
     * @nullable
     */
  personality?: string | null;
  /**
     * JSON-encoded preferences (budget, travel style, dietary, etc.)
     * @nullable
     */
  preferences?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  email?: string;
  phone?: string;
  location?: string;
  bio?: string;
  personality?: string;
  preferences?: string;
}

export interface Event {
  id: number;
  title: string;
  /** wedding, birthday, funeral, vacation, reunion, corporate, other */
  type: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  isInternational: boolean;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  guestCount?: number | null;
  /** planning, confirmed, completed, cancelled */
  status: string;
  /** @nullable */
  coverImage?: string | null;
  createdAt: string;
}

export interface EventInput {
  title: string;
  type: string;
  description?: string;
  location?: string;
  isInternational?: boolean;
  startDate?: string;
  endDate?: string;
  budget?: string;
  guestCount?: number;
}

export interface EventUpdate {
  title?: string;
  type?: string;
  description?: string;
  location?: string;
  isInternational?: boolean;
  startDate?: string;
  endDate?: string;
  budget?: string;
  guestCount?: number;
  status?: string;
  coverImage?: string;
}

export interface EventSummary {
  eventId: number;
  title: string;
  guestCount: number;
  confirmedGuests: number;
  savedSuggestions: number;
  sessionCount: number;
  inviteCount: number;
  completionPercent: number;
  nextSteps: string[];
}

export interface Dashboard {
  totalEvents: number;
  upcomingEvents: number;
  totalGuests: number;
  events: Event[];
}

export interface PlanningSession {
  id: number;
  eventId: number;
  title: string;
  /**
     * venue, activities, guests, dates, general
     * @nullable
     */
  focus?: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  title: string;
  focus?: string;
}

export interface Message {
  id: number;
  sessionId: number;
  /** user or assistant */
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface Guest {
  id: number;
  eventId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** pending, confirmed, declined */
  rsvpStatus: string;
  /**
     * JSON-encoded personality traits
     * @nullable
     */
  personality?: string | null;
  /** @nullable */
  dietaryNeeds?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface GuestInput {
  name: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  personality?: string;
  dietaryNeeds?: string;
  notes?: string;
}

export interface GuestUpdate {
  name?: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  rsvpStatus?: string;
  personality?: string;
  dietaryNeeds?: string;
  notes?: string;
}

export interface GuestPair {
  guest1: Guest;
  guest2: Guest;
  compatibilityScore: number;
  reason: string;
}

export interface SeatingGroup {
  tableName: string;
  guests: Guest[];
  vibe: string;
}

export interface GuestPairings {
  eventId: number;
  roommates: GuestPair[];
  travelBuddies: GuestPair[];
  seatingGroups: SeatingGroup[];
  reasoning: string;
}

export interface Suggestion {
  id: number;
  eventId: number;
  /** venue, activity, accommodation, dining, entertainment */
  type: string;
  name: string;
  description: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  estimatedCost?: string | null;
  /** @nullable */
  bestFor?: string | null;
  /**
     * JSON array of highlight strings
     * @nullable
     */
  highlights?: string | null;
  isSaved: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  createdAt: string;
}

export interface SuggestionGenerateInput {
  /** venue, activities, accommodation, dining, entertainment, all */
  focus: string;
  /** Additional preferences or constraints */
  preferences?: string;
  /** Number of suggestions to generate */
  count?: number;
}

export interface SuggestionUpdate {
  isSaved?: boolean;
  notes?: string;
}

export interface Invite {
  id: number;
  eventId: number;
  title: string;
  message: string;
  /** formal, casual, playful, elegant */
  style: string;
  whatsappText: string;
  whatsappLink: string;
  /**
     * JSON array of guest IDs this invite was sent to
     * @nullable
     */
  guestIds?: string | null;
  createdAt: string;
}

export interface InviteGenerateInput {
  /** formal, casual, playful, elegant */
  style: string;
  additionalContext?: string;
  guestIds?: number[];
}