# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Celebration Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new celebration event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event details
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{eventId}/summary:
    get:
      operationId: getEventSummary
      tags: [events]
      summary: Get event planning summary
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSummary"
...
  /events/{eventId}/guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: addGuest
      tags: [guests]
      summary: Add a guest to an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Added guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /events/{eventId}/guests/{guestId}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update guest details
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Remove a guest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: guestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /events/{eventId}/guests/pairings:
    get:
      operationId: getGuestPairings
      tags: [guests]
      summary: Get AI-suggested roommate, travel buddy, and seating pairings
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pairing suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestPairings"
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          description: "wedding, birthday, funeral, vacation, reunion, corporate, other"
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isInternational:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        budget:
          type: ["string", "null"]
        guestCount:
          type: ["integer", "null"]
        status:
          type: string
          description: "planning, confirmed, completed, cancelled"
        coverImage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, status, isInternational, createdAt]
    EventInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        location:
          type: string
        isInternational:
          type: boolean
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: string
        guestCount:
          type: integer
      required: [title, type]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        location:
          type: string
        isInternational:
          type: boolean
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: string
        guestCount:
          type: integer
        status:
          type: string
        coverImage:
          type: string
    EventSummary:
      type: object
      properties:
        eventId:
          type: integer
        title:
          type: string
        guestCount:
          type: integer
        confirmedGuests:
          type: integer
        savedSuggestions:
          type: integer
        sessionCount:
          type: integer
        inviteCount:
          type: integer
        completionPercent:
          type: integer
        nextSteps:
          type: array
          items:
            type: string
      required: [eventId, title, guestCount, confirmedGuests, savedSuggestions, sessionCount, inviteCount, completionPercent, nextSteps]
...
    Guest:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        rsvpStatus:
          type: string
          description: "pending, confirmed, declined"
        personality:
          type: ["string", "null"]
          description: JSON-encoded personality traits
        dietaryNeeds:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, name, rsvpStatus, createdAt]
    GuestInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        personality:
          type: string
        dietaryNeeds:
          type: string
        notes:
          type: string
      required: [name]
    GuestUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        rsvpStatus:
          type: string
        personality:
          type: string
        dietaryNeeds:
          type: string
        notes:
          type: string
    GuestPairings:
      type: object
      properties:
        eventId:
          type: integer
        roommates:
          type: array
          items:
            $ref: "#/components/schemas/GuestPair"
        travelBuddies:
          type: array
          items:
            $ref: "#/components/schemas/GuestPair"
        seatingGroups:
          type: array
          items:
            $ref: "#/components/schemas/SeatingGroup"
        reasoning:
          type: string
      required: [eventId, roommates, travelBuddies, seatingGroups, reasoning]
    GuestPair:
      type: object
      properties:
        guest1:
          $ref: "#/components/schemas/Guest"
        guest2:
          $ref: "#/components/schemas/Guest"
        compatibilityScore:
          type: integer
        reason:
          type: string
      required: [guest1, guest2, compatibilityScore, reason]
    SeatingGroup:
      type: object
      properties:
        tableName:
          type: string
        guests:
          type: array
          items:
            $ref: "#/components/schemas/Guest"
        vibe:
          type: string
      required: [tableName, guests, vibe]
...