# lib/api-zod/src/generated/types/eventSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface EventSummary {
  eventId: number;
  title: string;
  guestCount: number;
  confirmedGuests: number;
  savedSuggestions: number;
  sessionCount: number;
  inviteCount: number;
  completionPercent: number;
  nextSteps: string[];
}