# lib/api-zod/src/generated/types/guestInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestInput {
  name: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  personality?: string;
  dietaryNeeds?: string;
  notes?: string;
}