# lib/api-zod/src/generated/types/guestPair.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { Guest } from './guest';

export interface GuestPair {
  guest1: Guest;
  guest2: Guest;
  compatibilityScore: number;
  reason: string;
}