# lib/api-zod/src/generated/types/guestPairings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestPair } from './guestPair';
import type { SeatingGroup } from './seatingGroup';

export interface GuestPairings {
  eventId: number;
  roommates: GuestPair[];
  travelBuddies: GuestPair[];
  seatingGroups: SeatingGroup[];
  reasoning: string;
}