# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Celebration Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestUpdate {
  name?: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  rsvpStatus?: string;
  personality?: string;
  dietaryNeeds?: string;
  notes?: string;
}